WP-5.2: swap Splash.Silk onto the generated GL bindings
Silk.NET.OpenGL is gone from the render path. 22 files, +41/-56.

The swap itself is what WP-5.0 predicted: a `using` change. 21 files needed only
`using Silk.NET.OpenGL;` -> `using Karawan.Graphics.OpenGL;`, and every call site
compiled unchanged, which was the entire premise of generating names that match.

FOUR PLACES NEEDED MORE, and they are worth naming because WP-5.0's "exactly 0
changed lines" was measured on a five-entry-point sample and is not quite true at
full scale:

  PreviewHelper       DelegateProcContext deleted. It existed ONLY to satisfy
                      Silk's GetApi(INativeContext); the generated binding takes
                      the Func<string,IntPtr> directly, which is what Aihao was
                      already passing.
  ImGui/Texture.cs    TextureCoordinate's initialisers now cast. The generated
                      TextureParameterName is uint-backed because gl.xml calls
                      these GLenum; Silk's was int-backed. This is exactly the
                      "signedness only" category the WP-5.1 differ reports, and
                      it is the ONE place in Splash.Silk where the swap needed a
                      real source change.
  ImGui/Controller.cs a dead `private Shader _shader;` removed - declared, never
                      assigned, never read. It had been resolving to Silk's
                      Shader HANDLE struct, which the generated binding does not
                      define because no call site uses it.
  Platform.cs         _gl?.Dispose() removed. The generated binding is not
                      IDisposable and has nothing to dispose: a resolver delegate
                      and lazily-created P/Invoke delegates, no GL objects, no
                      unmanaged allocation. Silk's GL implemented IDisposable for
                      its own loader bookkeeping.

Silk.NET.OpenGL.Extensions.EXT was dropped too - it had ZERO usages, so it had
been a dead reference for some time.

The #if GLES / #elif LEGACY import branches are deliberately untouched. They are
not compiled, and swapping them would claim a portability nobody has tested.

VERIFIED

  dotnet build Karawan.sln                     0 errors
  unit tests                                   234/234
  wp-5.1 differ (signature parity vs Silk)     exit 0
  Silk.NET.OpenGL in Karawan's package graph   ABSENT, transitively too
  game run, autoLogin=new, ~90 s               4911 frames, up to 141 mesh
                                               batches, ZERO GL errors and zero
                                               exceptions in the log

What remains of Silk anywhere is Silk.NET.Assimp (Phase 4, pinned per N5/N8) with
Silk.NET.Core and .Maths riding in behind it.

NOT VERIFIED: that it LOOKS right. Frames are produced with plausible content and
no GL errors, which is strong but is not the same claim - GATE-E and a look at
the screen are still the owner's. A wrong enum in GlStateSaver or SilkRenderState
fails silently rather than loudly, which is precisely why the plan wanted eyes on
it.

+++ Splash.Silk/ImGui/Texture.cs
@@ -1,4 +1,4 @@
-using Silk.NET.OpenGL;
+using Karawan.Graphics.OpenGL;
 
 namespace Splash.Silk.ImGui;
 
@@ -9,16 +9,20 @@ using System;
 #if GLES
 using Silk.NET.OpenGLES;
 #elif GL
-using Silk.NET.OpenGL;
+using Karawan.Graphics.OpenGL;
 #elif LEGACY
 using Silk.NET.OpenGL.Legacy;
 #endif
 
 public enum TextureCoordinate
 {
-    S = TextureParameterName.TextureWrapS,
-    T = TextureParameterName.TextureWrapT,
-    R = TextureParameterName.TextureWrapR
+    // Cast because the generated TextureParameterName is uint-backed (gl.xml calls these
+    // GLenum) while Silk's was int-backed. Same values, same ABI - this is the
+    // "signedness only" category the WP-5.1 differ reports, and it is the one place in
+    // Splash.Silk where the swap needed a source change at all.
+    S = (int)TextureParameterName.TextureWrapS,
+    T = (int)TextureParameterName.TextureWrapT,
+    R = (int)TextureParameterName.TextureWrapR
 }
 
 class Texture : IDisposable
